+++ resources/sql/queries.sql
@@ -54,3 +54,15 @@ SELECT * from problems order by num
 -- :name delete-problems-all! :! :n
 -- :doc delete all from problems table
 DELETE FROM problems
+
+-- :name create-answer! :! :n
+-- :doc creates a new problem record
+INSERT INTO answers
+(login, num, answer, md5)
+VALUES (:login, :num, :answer, md5)
+
+-- :name get-answer :? :1
+-- :doc retrieves an answer to `num` from user `login`
+SELECT * FROM answers
+WHERE num = :num and login = :login
+